fix: 🐛 auth callback issue
