+++ package.json
@@ -46,7 +46,7 @@
     "astro": "^5.1.7",
     "astro-breadcrumbs": "^3.3.1",
     "astro-font": "^0.1.81",
-    "better-auth": "^1.1.11",
+    "better-auth": "^1.1.15",
     "class-variance-authority": "^0.7.1",
     "contentful": "^11.3.4",
     "contentful-management": "^11.40.3",
@@ -64,14 +64,14 @@
     "ts-pattern": "^5.6.0"
   },
   "devDependencies": {
+    "@better-auth/cli": "^1.1.14",
     "@eslint/js": "^9.19.0",
     "@tailwindcss/postcss": "^4.0.0",
     "@tailwindcss/vite": "^4.0.0",
     "@types/eslint": "^9.6.1",
-    "@better-auth/cli": "^1.1.14",
     "@types/html-to-text": "^9.0.4",
-    "@typescript-eslint/parser": "^8.22.0",
     "@types/json-patch": "^0.0.33",
+    "@typescript-eslint/parser": "^8.22.0",
     "autoprefixer": "^10.4.20",
     "cf-content-types-generator": "^2.15.5",
     "contentful-merge": "^2.6.5",
@@ -92,9 +92,9 @@
     "prettier-plugin-tailwindcss": "^0.6.11",
     "react-compiler-runtime": "^19.0.0-beta-b2e8e9c-20241220",
     "react-email": "3.0.6",
+    "tsx": "^4.19.2",
     "typescript": "^5.7.3",
     "typescript-eslint": "^8.22.0",
-    "tsx": "^4.19.2",
     "vitest": "^3.0.4",
     "zx": "^8.3.0"
   },
