+++ src/lib/email/types.ts
@@ -1,0 +1,5 @@
+export type Email = {
+  to: string;
+  subject: string;
+  html: string;
+};
